Holmes/ApproachImp/EvidenceGather/requirements.txt
asn1crypto==0.22.0
backcall==0.2.0
backports-datetime-fromisoformat==2.0.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
certifi==2016.2.28
cffi==1.10.0
charset-normalizer==3.0.1
colorama==0.3.9
courlan==0.8.3
cycler==0.11.0
dateparser==1.1.3
decorator==4.4.2
Deprecated==1.2.12
graphviz==0.19.1
htmldate==1.4.1
idna==2.6
importlib-resources==5.4.0
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.10.2
joblib==1.1.1
JPype1==1.3.0
jusText==3.0.0
keras==2.10.0
kiwisolver==1.3.1
LambdaRankNN==0.1.1
langcodes==3.3.0
lxml==4.9.2
matplotlib==3.3.4
networkx==2.5.1
nltk==3.2.4
numpy==1.19.5
packaging==16.8
path.py==10.3.1
pickleshare==0.7.4
Pillow==8.4.0
prompt-toolkit==3.0.20
pycparser==2.18
pydot==1.4.1
PyGithub==1.54.1
Pygments==2.2.0
PyJWT==1.7.1
pymongo==4.1.1
pyOpenSSL==17.0.0
pyparsing==2.2.0
PySocks==1.6.6
python-dateutil==2.8.2
pytz==2017.2
pytz-deprecation-shim==0.1.0.post0
redis==3.5.3
regex==2017.4.5
requests==2.14.2
scikit-learn==0.24.2
scipy==1.5.4
selenium==3.14.0
six==1.10.0
sklearn==0.0.post2
soupsieve==2.3.1
threadpoolctl==3.1.0
tld==0.12.7
trafilatura==1.4.1
traitlets==4.3.2
typing-extensions==4.1.1
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
validators==0.18.2
version-information==1.0.4
wcwidth==0.1.7
win-inet-pton==1.0.1
wincertstore==0.2
wrapt==1.10.11
xgboost==1.5.2
xlrd==1.1.0
zipp==3.6.0
